refactor: remove unused Loading components, keep only SkeletonCard
